Split up DataType
