+++ src/Parsers/CreateDefinitions.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace PhpMyAdmin\SqlParser\Parsers;
 
 use PhpMyAdmin\SqlParser\Components\CreateDefinition;
-use PhpMyAdmin\SqlParser\Components\DataType;
 use PhpMyAdmin\SqlParser\Components\Key;
 use PhpMyAdmin\SqlParser\Components\OptionsArray;
 use PhpMyAdmin\SqlParser\Parseable;
@@ -195,7 +194,7 @@ final class CreateDefinitions implements Parseable
                     return $ret;
                 }
             } elseif ($state === 2) {
-                $expr->type = DataType::parse($parser, $list);
+                $expr->type = DataTypes::parse($parser, $list);
                 $state = 3;
             } elseif ($state === 3) {
                 $expr->options = OptionsArray::parse($parser, $list, self::FIELD_OPTIONS);
